cleanup(storage): fix GCS tracing test on macOS

+++ google/cloud/storage/internal/async/connection_tracing_test.cc
@@ -675,7 +675,7 @@ TEST(ConnectionTracing, OpenSuccessWithInitialReadRanges) {
   EXPECT_THAT(
       spans, ElementsAre(
                  AllOf(SpanNamed("storage::AsyncConnection::Open"),
-                       SpanHasAttributes(OTelAttribute<std::size_t>(
+                       SpanHasAttributes(OTelAttribute<std::uint64_t>(
                            "gl-cpp.initial-read-ranges.ranges-count", 2)),
                        SpanWithStatus(opentelemetry::trace::StatusCode::kOk))));
 }
